Revert "MMAPS: Update recast 1.st part"
This reverts commit f8516ed1360f735dc0c56718d0e181b30d8c4ef1.

+++ src/tools/mesh_extractor/MPQ.cpp
@@ -77,7 +77,7 @@ size_t MPQFile::Read(void* dest, size_t bytes)
         return 0;
 
     size_t rpos = pointer + bytes;
-    if (rpos > size_t(size)) {
+    if (rpos > size) {
         bytes = size - pointer;
         eof = true;
     }
